requirements.txt
Jinja2==3.1.4
MarkupSafe==2.1.5
PyYAML==6.0.2
Pygments==2.18.0
SQLAlchemy==2.0.32
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==2.9.2
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi>=0.68.0
fastapi-cli==0.0.5
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
jaraco.classes==3.4.0
keyring==24.3.1
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.4.0
pip==24.1.2
polyline==2.0.2
py==1.11.0
pybind11==2.9.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
rich==13.7.1
setuptools==72.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn>=0.15.0
uvloop==0.19.0
watchfiles==0.23.0
websockets==12.0
wheel==0.43.0
zipp==3.20.0
APScheduler==3.10.4
psycopg2-binary==2.9.9